refactor: move Eidoverse world-projection schemas into eidoverseValidation.js (#5698)
validation.js is the repo's second-hottest file and its own README calls it a
catch-all. The Eidoverse world projection was its largest un-migrated tenant —
306 contiguous lines under a self-delimiting banner, 20 module-private helper
schemas, and nothing in the block referencing a symbol declared elsewhere in
the file — so every Eidoverse recipe edit collided with unrelated app/provider/
backup schema churn.

Pure move, verbatim: the extracted body is byte-identical to the deleted lines.
The new module imports only zod (importing back through validation.js would
TDZ — ESM hoists the `export * from` line), and validation.js re-exports it
from the transitional #1151 block, so no consumer's import specifier changes.

Also adds the barrel namespace export + README row the server/lib maintenance
rule requires, boundary tests for each public schema (source-key allowlist,
asset-path escape guard, 8KB augment-argument cap), and one case pinning that
eidoverseWorldConfigPatchSchema still resolves through validation.js.
